# public/static/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "lanchonete API",
    "version" : "1.0.2"
  },
...
  "paths" : {
    "/Produtos" : {
      "post" : {
        "tags" : [ "Produtos" ],
        "summary" : "Insere um produto na base de dados",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProdutoDto"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Produto inserido na base de dados",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateProdutoDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Produtos" ],
        "summary" : "Deleta um produto da base de dados",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeleteProdutoDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Produto deletado na base de dados",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeleteProdutoDto"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Produtos" ],
        "summary" : "Edita um produto da base de dados",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditProdutoDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Produto editado na base de dados",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EditProdutoDto"
                }
              }
            }
          }
        }
      }
    },
    "/produtos/tipo/{codigoTipoProduto}" : {
      "get" : {
        "tags" : [ "Produtos" ],
        "summary" : "Lista produtos por tipo",
        "parameters" : [ {
          "name" : "codigoTipoProduto",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Produto"
                  }
                }
              }
            }
          },
          "404" : { }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Produto" : {
        "type" : "object",
        "properties" : {
          "codigoTipoProduto" : {
            "type" : "integer"
          },
          "codigoProduto" : {
            "type" : "integer"
          },
          "nomeProduto" : {
            "type" : "string"
          },
          "nomeFunciAlter" : {
            "type" : "string"
          },
          "precoProduto" : {
            "type" : "number"
          },
          "tsAlter" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "CreateProdutoDto" : {
        "required" : [ "nomeProduto", "nomeFunciAlter" ],
        "type" : "object",
        "properties" : {
          "codigoTipoProduto" : {
            "type" : "integer"
          },
          "nomeProduto" : {
            "minLength" : 1,
            "type" : "string"
          },
          "nomeFunciAlter" : {
            "minLength" : 1,
            "type" : "string"
          },
          "precoProduto" : {
            "minimum" : 0,
            "exclusiveMinimum" : true,
            "type" : "number"
          }
        }
      },
...
      "DeleteProdutoDto" : {
        "type" : "object",
        "properties" : {
          "codigoProduto" : {
            "type" : "integer"
          },
          "codigoTipoProduto" : {
            "type" : "integer"
          }
        }
      },
      "EditProdutoDto" : {
        "type" : "object",
        "properties" : {
          "codigoTipoProduto" : {
            "type" : "integer"
          },
          "codigoProduto" : {
            "type" : "integer"
          },
          "nomeProduto" : {
            "type" : "string"
          },
          "nomeFunciAlter" : {
            "type" : "string"
          },
          "precoProduto" : {
            "type" : "number"
          }
        }
      },
...
      "Instant" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T16:15:50Z"
      },
...
    }
  }
}