# src/main/java/com/fiap/lanchonete/domain/pojo/ProdutoDto.java
package com.fiap.lanchonete.domain.pojo;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class ProdutoDto {
    private Short codigoTipoProduto;
    private Short codigoProduto;
    private Integer quantidadeProduto;
}